+++ CHANGELOG.md
@@ -4,6 +4,8 @@ v1.7.0
  - Finder增加GetValues方法,返回Finder封装的values值
  - zormErrorSQLValues 使用json格式输出
  - BindContextOnlyUpdateCols 指定仅更新的数据库字段,只对Update方法有效.cols是数据库列名切片
+ - BindContextMustUpdateCols 指定必须更新的数据库字段,只对UpdateNotZeroValue方法有效.cols是数据库列名切片
+ - BindContextDefaultValue 替代 BindContextMustUpdate函数
  - 完善文档,注释
 
 v1.6.9